move internal/codeintel/lsif/lsif => lib/codeintel/lsif/conversion
